Decompose pre course assessment form and element renderer
