+++ apps/platform/src/lib/infrastructure/client/utils/transform-answers.ts
@@ -5,6 +5,7 @@ import {
     SingleChoiceElement,
     MultiCheckElement,
     OneOutOfThreeElement,
+    LessonElement,
 } from '@maany_shr/e-class-ui-kit';
 
 function transformTextInputAnswer(
@@ -104,7 +105,7 @@ const answerTransformers = {
 } as const;
 
 export function transformFormAnswers(
-    formValues: Record<string, FormElement>,
+    formValues: Record<string, LessonElement>,
 ): useCaseModels.TAnswer[] {
     const answers: useCaseModels.TAnswer[] = [];
