+++ packages/ui-kit/lib/components/lesson/submission-renderer.tsx
@@ -1,15 +1,15 @@
 import { isLocalAware } from "@maany_shr/e-class-translations";
-import { FormElement } from "./types";
-import { formElements } from "./form-element-core";
+import { LessonElement } from "./types";
+import { lessonElements } from "./element-core";
 
 interface SubmissionRendererProps extends isLocalAware {
-    elements: FormElement[];
+    elements: LessonElement[];
 }
 
 export function SubmissionElementsRenderer({elements, locale}: SubmissionRendererProps) {
     return <div className="flex flex-col gap-4">
         {elements.map((elementInstance) => {
-            const Element = formElements[elementInstance.type].submissionComponent;
+            const Element = lessonElements[elementInstance.type].submissionComponent;
             return (
                 <div key={elementInstance.id.toString()} className="flex flex-col gap-2 items-start">
                     <Element
