+++ packages/ui-kit/stories/form-element-renderer.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from "@storybook/react";
-import { FormElementRenderer } from "../lib/components/pre-assessment/form-renderer";
+import { FormElementRenderer } from "../lib/components/lesson/form-renderer";
 import { FormElementType, FormElement } from "../lib/components/pre-assessment/types";
 import { slateifySerialize } from "../lib/components/rich-text-element/serializer";
 const meta: Meta<typeof FormElementRenderer> = {
